+++ src/main/java/com/github/offsetmonkey538/baguette/config/BaguetteConfig.java
@@ -3,25 +3,33 @@ package com.github.offsetmonkey538.baguette.config;
 import java.io.Serializable;
 
 public final class BaguetteConfig implements Serializable {
-    private int tntBaguetteExplosionStrength =            6;
+    private int   tntBaguetteExplosionStrength =             6;
 
-    private int chargedTntBaguetteExplosionStrength =     12;
+    private int   chargedTntBaguetteExplosionStrength =      12;
 
-    private int waterBaguetteHungerDurationTicks =        1200;
-    private int waterBaguetteHungerAmplifier =            0;
-    private float waterBaguetteHungerChance =             100;
+    private int   waterBaguetteHungerDurationTicks =         1200;
+    private int   waterBaguetteHungerAmplifier =             0;
+    private float waterBaguetteHungerChance =                100;
 
-    private int waterBaguetteNauseaDurationTicks =        600;
-    private int waterBaguetteNauseaAmplifier =            0;
-    private float waterBaguetteNauseaChance =             100;
+    private int   waterBaguetteNauseaDurationTicks =         600;
+    private int   waterBaguetteNauseaAmplifier =             0;
+    private float waterBaguetteNauseaChance =                100;
 
-    private int waterBaguetteWaterBreathingDurationTicks =        3600;
-    private int waterBaguetteWaterBreathingAmplifier =            0;
-    private float waterBaguetteWaterBreathingChance =             100;
+    private int   waterBaguetteWaterBreathingDurationTicks = 3600;
+    private int   waterBaguetteWaterBreathingAmplifier =     0;
+    private float waterBaguetteWaterBreathingChance =        100;
 
-    private int fireBaguetteFireResistanceDurationTicks = 3600;
-    private int fireBaguetteFireResistanceAmplifier =     0;
-    private float fireBaguetteFireResistanceChance =      100;
+    private int   fireBaguetteFireResistanceDurationTicks =  3600;
+    private int   fireBaguetteFireResistanceAmplifier =      0;
+    private float fireBaguetteFireResistanceChance =         100;
+
+    private int   airBaguetteSlowFallingDurationTicks =      3600;
+    private int   airBaguetteSlowFallingAmplifier =          0;
+    private float airBaguetteSlowFallingChance =             100;
+
+    private int   airBaguetteJumpBoostDurationTicks =        3600;
+    private int   airBaguetteJumpBoostAmplifier =            0;
+    private float airBaguetteJumpBoostChance =               100;
 
     public BaguetteConfig() {
 
@@ -73,7 +81,24 @@ public final class BaguetteConfig implements Serializable {
     public void setWaterBaguetteWaterBreathingChance(float waterBaguetteWaterBreathingChance) {
         this.waterBaguetteWaterBreathingChance = waterBaguetteWaterBreathingChance;
     }
-
+    public void setAirBaguetteSlowFallingDurationTicks(int airBaguetteSlowFallingDurationTicks) {
+        this.airBaguetteSlowFallingDurationTicks = airBaguetteSlowFallingDurationTicks;
+    }
+    public void setAirBaguetteSlowFallingAmplifier(int airBaguetteSlowFallingAmplifier) {
+        this.airBaguetteSlowFallingAmplifier = airBaguetteSlowFallingAmplifier;
+    }
+    public void setAirBaguetteSlowFallingChance(float airBaguetteSlowFallingChance) {
+        this.airBaguetteSlowFallingChance = airBaguetteSlowFallingChance;
+    }
+    public void setAirBaguetteJumpBoostDurationTicks(int airBaguetteJumpBoostDurationTicks) {
+        this.airBaguetteJumpBoostDurationTicks = airBaguetteJumpBoostDurationTicks;
+    }
+    public void setAirBaguetteJumpBoostAmplifier(int airBaguetteJumpBoostAmplifier) {
+        this.airBaguetteJumpBoostAmplifier = airBaguetteJumpBoostAmplifier;
+    }
+    public void setAirBaguetteJumpBoostChance(float airBaguetteJumpBoostChance) {
+        this.airBaguetteJumpBoostChance = airBaguetteJumpBoostChance;
+    }
 
     //
     // Getters
@@ -120,5 +145,23 @@ public final class BaguetteConfig implements Serializable {
     }
     public float getWaterBaguetteWaterBreathingChance() {
         return waterBaguetteWaterBreathingChance;
+    }
+    public int getAirBaguetteSlowFallingDurationTicks() {
+        return airBaguetteSlowFallingDurationTicks;
+    }
+    public int getAirBaguetteSlowFallingAmplifier() {
+        return airBaguetteSlowFallingAmplifier;
+    }
+    public float getAirBaguetteSlowFallingChance() {
+        return airBaguetteSlowFallingChance;
+    }
+    public int getAirBaguetteJumpBoostDurationTicks() {
+        return airBaguetteJumpBoostDurationTicks;
+    }
+    public int getAirBaguetteJumpBoostAmplifier() {
+        return airBaguetteJumpBoostAmplifier;
+    }
+    public float getAirBaguetteJumpBoostChance() {
+        return airBaguetteJumpBoostChance;
     }
 }