+++ src/main/java/com/github/offsetmonkey538/baguette/item/ModItems.java
@@ -2,6 +2,7 @@ package com.github.offsetmonkey538.baguette.item;
 
 import com.github.offsetmonkey538.baguette.BaguetteMain;
 import com.github.offsetmonkey538.baguette.item.baguettes.*;
+import com.github.offsetmonkey538.baguette.item.baguettes.elements.AirBaguette;
 import com.github.offsetmonkey538.baguette.item.baguettes.elements.FireBaguette;
 import com.github.offsetmonkey538.baguette.item.baguettes.elements.WaterBaguette;
 import com.github.offsetmonkey538.baguette.item.tools.LongBaguetteSword;
@@ -24,6 +25,7 @@ public final class ModItems {
     public static final Item WATER_BAGUETTE       = addItem(new WaterBaguette(      new Item.Settings().group(ItemGroup.FOOD).food(ModFoodComponents.WATER_BAGUETTE)),       "water_baguette");
     public static final Item BIRTHDAY_BAGUETTE    = addItem(new BirthdayBaguette(   new Item.Settings().group(ItemGroup.FOOD).food(ModFoodComponents.BIRTHDAY_BAGUETTE)),    "birthday_baguette");
     public static final Item FIRE_BAGUETTE        = addItem(new FireBaguette(       new Item.Settings().group(ItemGroup.FOOD).food(ModFoodComponents.FIRE_BAGUETTE)),        "fire_baguette");
+    public static final Item AIR_BAGUETTE         = addItem(new AirBaguette(        new Item.Settings().group(ItemGroup.FOOD).food(ModFoodComponents.AIR_BAGUETTE)),         "air_baguette");
     public static final Item LONG_BAGUETTE_SWORD  = addItem(new LongBaguetteSword(  new Item.Settings().group(ItemGroup.COMBAT)),                                            "long_baguette_sword");
     public static final Item EASTEREGG_BAGUETTE   = addItem(new EastereggBaguette(  new Item.Settings().group(ItemGroup.MISC)),                                              "easteregg_baguette");