+++ src/main/java/com/github/offsetmonkey538/baguette/item/baguettes/elements/AirBaguette.java
@@ -1,0 +1,10 @@
+package com.github.offsetmonkey538.baguette.item.baguettes.elements;
+
+import com.github.offsetmonkey538.baguette.item.baguettes.Baguette;
+
+public class AirBaguette extends Baguette {
+
+    public AirBaguette(Settings settings) {
+        super(settings);
+    }
+}