+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Added
 - Fire Baguette: Spawns a ring of fire around you and gives you the Fire Resistance effect.
+- Air Baguette: Gives you the Slow Falling effect and the Jump Boost effect.
 
 ## Changed
 - Water Baguette now gives Water Breathing effect.